+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `Full Impersonation` Organization Setting, this will allow a Impersonator to also Switch between the User's Organization/Cost Center
 
+## [1.35.2] - 2024-10-15
+### Added
+
+- Add new tab called 'settings' at organization details
+
 ## [1.35.1] - 2024-10-01
 
 ## [1.35.0] - 2024-09-10